+++ pages/search.tsx
@@ -1,0 +1,11 @@
+import Seo from 'components/Layout/Seo/Seo';
+import SearchWrapper from 'components/Search/MainSearchWrapper';
+
+const SearchResidentPage = (): React.ReactElement => (
+  <div>
+    <Seo title="Search" />
+    <SearchWrapper type="people" />
+  </div>
+);
+
+export default SearchResidentPage;
