+++ stylesheets/all.scss
@@ -134,3 +134,7 @@ $screen-sm-min: 576px;
   border-left-width: 5px;
   background-color: #f8ecea;
 }
+
+.govuk-heading-l {
+  font-weight: 600;
+}
